Avoid preprend src in the imports
