+++ src/main.py
@@ -7,8 +7,8 @@ from typing import List, Union, Tuple
 import fitz
 import pandas as pd
 
-from src.workbook import write_style_frame, Sentence
-from src.predict_strategy import (
+from workbook import write_style_frame, Sentence
+from predict_strategy import (
     STRATEGY_FLAIR,
     FaceImagePredictor,
     Span,
